+++ tests/test_protocol/test_responseprocessor_exceptions.py
@@ -7,7 +7,7 @@ class MockResponse:
     def __init__(self, data):
         self.data = data
 
-    def model_dump(self):
+    def model_dump(self, *args, **kwargs):
         return self.data
