+++ tests/test_protocol/test_responses_v2.py
@@ -221,7 +221,7 @@ class MockResponse:
     def __init__(self, data):
         self.data = data
 
-    def model_dump(self):
+    def model_dump(self, *args, **kwargs):
         return self.data
